extract helper into inner class, remove abstract class
