+++ app/com/arpnetworking/metrics/portal/reports/impl/DatabaseReportExecutionRepository.java
@@ -19,9 +19,7 @@ import com.arpnetworking.metrics.portal.reports.ReportExecutionRepository;
 import com.arpnetworking.metrics.portal.scheduling.JobExecutionRepository;
 import com.arpnetworking.steno.Logger;
 import com.arpnetworking.steno.LoggerFactory;
-import com.google.common.base.Throwables;
 import io.ebean.EbeanServer;
-import io.ebean.Transaction;
 import models.ebean.ReportExecution;
 import models.internal.Organization;
 import models.internal.reports.Report;
@@ -32,12 +30,9 @@ import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Consumer;
-import javax.annotation.Nullable;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.persistence.EntityNotFoundException;
-import javax.persistence.PersistenceException;
 
 /**
  * Implementation of {@link JobExecutionRepository} for {@link Report} jobs using a SQL database.
@@ -50,6 +45,7 @@ public final class DatabaseReportExecutionRepository implements ReportExecutionR
 
     private final AtomicBoolean _isOpen = new AtomicBoolean(false);
     private final EbeanServer _ebeanServer;
+    private final DatabaseExecutionRepositoryHelper<Report.Result, ReportExecution> _helper;
 
     /**
      * Public constructor.
@@ -59,6 +55,41 @@ public final class DatabaseReportExecutionRepository implements ReportExecutionR
     @Inject
     public DatabaseReportExecutionRepository(@Named("metrics_portal") final EbeanServer ebeanServer) {
         _ebeanServer = ebeanServer;
+        _helper = new DatabaseExecutionRepositoryHelper<>(LOGGER, _ebeanServer, this::reportExecutionSupplier);
+
+    }
+
+    private ReportExecution reportExecutionSupplier(
+            final UUID jobId,
+            final Organization organization,
+            final Instant scheduled
+    ) {
+        final Optional<models.ebean.Report> report = models.ebean.Organization.findByOrganization(_ebeanServer, organization)
+                .flatMap(beanOrg -> models.ebean.Report.findByUUID(
+                        _ebeanServer,
+                        beanOrg,
+                        jobId
+                ));
+        if (!report.isPresent()) {
+            final String message = String.format(
+                    "Could not find report with uuid=%s, organization.uuid=%s",
+                    jobId,
+                    organization.getId()
+            );
+            throw new EntityNotFoundException(message);
+        }
+
+        final Optional<ReportExecution> existingExecution = report.flatMap(r ->
+                _ebeanServer.createQuery(ReportExecution.class)
+                        .where()
+                        .eq("report", r)
+                        .eq("scheduled", scheduled)
+                        .findOneOrEmpty()
+        );
+        final ReportExecution newOrUpdatedExecution = existingExecution.orElse(new ReportExecution());
+        newOrUpdatedExecution.setReport(report.get());
+        newOrUpdatedExecution.setScheduled(scheduled);
+        return newOrUpdatedExecution;
     }
 
     @Override
@@ -74,7 +105,6 @@ public final class DatabaseReportExecutionRepository implements ReportExecutionR
         LOGGER.debug().setMessage("Closing DatabaseReportExecutionRepository").log();
         _isOpen.set(false);
     }
-
 
     /**
      * Get the most recently scheduled execution, if any.
@@ -97,7 +127,7 @@ public final class DatabaseReportExecutionRepository implements ReportExecutionR
                 .eq("report.organization.uuid", organization.getId())
                 .setMaxRows(1)
                 .findOneOrEmpty()
-                .map(this::toInternalModel);
+                .map(DatabaseExecutionRepositoryHelper::toInternalModel);
     }
 
     @Override
@@ -114,7 +144,7 @@ public final class DatabaseReportExecutionRepository implements ReportExecutionR
                 .setMaxRows(1)
                 .findOneOrEmpty();
         if (row.isPresent()) {
-            final JobExecution<Report.Result> execution = toInternalModel(row.get());
+            final JobExecution<Report.Result> execution = DatabaseExecutionRepositoryHelper.toInternalModel(row.get());
             if (execution instanceof JobExecution.Success) {
                 return Optional.of((JobExecution.Success<Report.Result>) execution);
             }
@@ -138,152 +168,25 @@ public final class DatabaseReportExecutionRepository implements ReportExecutionR
                 .in("state", ReportExecution.State.SUCCESS, ReportExecution.State.FAILURE)
                 .setMaxRows(1)
                 .findOneOrEmpty()
-                .map(this::toInternalModel);
+                .map(DatabaseExecutionRepositoryHelper::toInternalModel);
     }
 
     @Override
     public void jobStarted(final UUID reportId, final Organization organization, final Instant scheduled) {
         assertIsOpen();
-        updateExecutionState(
-                reportId,
-                organization,
-                scheduled,
-                ReportExecution.State.STARTED,
-                execution -> {
-                    execution.setStartedAt(Instant.now());
-                }
-        );
+        _helper.jobStarted(reportId, organization, scheduled);
     }
 
     @Override
     public void jobSucceeded(final UUID reportId, final Organization organization, final Instant scheduled, final Report.Result result) {
         assertIsOpen();
-        updateExecutionState(
-                reportId,
-                organization,
-                scheduled,
-                ReportExecution.State.SUCCESS,
-                execution -> {
-                    execution.setResult(result);
-                    execution.setCompletedAt(Instant.now());
-                }
-        );
+        _helper.jobSucceeded(reportId, organization, scheduled, result);
     }
 
     @Override
     public void jobFailed(final UUID reportId, final Organization organization, final Instant scheduled, final Throwable error) {
         assertIsOpen();
-        updateExecutionState(
-                reportId,
-                organization,
-                scheduled,
-                ReportExecution.State.FAILURE,
-                execution -> {
-                    execution.setError(Throwables.getStackTraceAsString(error));
-                    execution.setCompletedAt(Instant.now());
-                }
-        );
-    }
-
-    private void updateExecutionState(
-            final UUID reportId,
-            final Organization organization,
-            final Instant scheduled,
-            final ReportExecution.State state,
-            final Consumer<ReportExecution> update
-    ) {
-        LOGGER.debug()
-                .setMessage("Upserting report execution")
-                .addData("report.uuid", reportId)
-                .addData("scheduled", scheduled)
-                .addData("state", state)
-                .log();
-        try (Transaction transaction = _ebeanServer.beginTransaction()) {
-            final Optional<models.ebean.Report> report = models.ebean.Organization.findByOrganization(_ebeanServer, organization)
-                    .flatMap(beanOrg -> models.ebean.Report.findByUUID(
-                            _ebeanServer,
-                            beanOrg,
-                            reportId
-                    ));
-            if (!report.isPresent()) {
-                final String message = String.format(
-                        "Could not find report with uuid=%s, organization.uuid=%s",
-                        reportId,
-                        organization.getId()
-                );
-                throw new EntityNotFoundException(message);
-            }
-
-            final Optional<ReportExecution> existingExecution = report.flatMap(r ->
-                    _ebeanServer.createQuery(ReportExecution.class)
-                            .where()
-                            .eq("report", r)
-                            .eq("scheduled", scheduled)
-                            .findOneOrEmpty()
-            );
-            final ReportExecution newOrUpdatedExecution = existingExecution.orElse(new ReportExecution());
-            newOrUpdatedExecution.setReport(report.get());
-            newOrUpdatedExecution.setScheduled(scheduled);
-
-            update.accept(newOrUpdatedExecution);
-            newOrUpdatedExecution.setState(state);
-
-            if (existingExecution.isPresent()) {
-                _ebeanServer.update(newOrUpdatedExecution);
-            } else {
-                _ebeanServer.save(newOrUpdatedExecution);
-            }
-
-            LOGGER.debug()
-                    .setMessage("Upserted report execution")
-                    .addData("report.uuid", reportId)
-                    .addData("scheduled", scheduled)
-                    .addData("state", state)
-                    .log();
-            transaction.commit();
-            // CHECKSTYLE.OFF: IllegalCatchCheck
-        } catch (final RuntimeException e) {
-            // CHECKSTYLE.ON: IllegalCatchCheck
-            LOGGER.error()
-                    .setMessage("Failed to upsert report executions")
-                    .addData("report.uuid", reportId)
-                    .addData("scheduled", scheduled)
-                    .addData("state", state)
-                    .setThrowable(e)
-                    .log();
-            throw new PersistenceException("Failed to upsert report executions", e);
-        }
-    }
-
-    private JobExecution<Report.Result> toInternalModel(final ReportExecution beanModel) {
-        final ReportExecution.State state = beanModel.getState();
-        switch (state) {
-            case STARTED:
-                return new JobExecution.Started.Builder<Report.Result>()
-                        .setJobId(beanModel.getReport().getUuid())
-                        .setScheduled(beanModel.getScheduled())
-                        .setStartedAt(beanModel.getStartedAt())
-                        .build();
-            case FAILURE:
-                @Nullable final Throwable throwable = beanModel.getError() == null ? null : new Throwable(beanModel.getError());
-                return new JobExecution.Failure.Builder<Report.Result>()
-                        .setJobId(beanModel.getReport().getUuid())
-                        .setScheduled(beanModel.getScheduled())
-                        .setStartedAt(beanModel.getStartedAt())
-                        .setCompletedAt(beanModel.getCompletedAt())
-                        .setError(throwable)
-                        .build();
-            case SUCCESS:
-                return new JobExecution.Success.Builder<Report.Result>()
-                        .setJobId(beanModel.getReport().getUuid())
-                        .setScheduled(beanModel.getScheduled())
-                        .setCompletedAt(beanModel.getCompletedAt())
-                        .setStartedAt(beanModel.getStartedAt())
-                        .setResult(beanModel.getResult())
-                        .build();
-            default:
-                throw new AssertionError("unexpected state: " + state);
-        }
+        _helper.jobFailed(reportId, organization, scheduled, error);
     }
 
     private void assertIsOpen() {
